refactor: Scale 이동
